Break out NM callback handler
